+++ PupasCorp/Models/Movimiento.cs
@@ -13,7 +13,7 @@ public partial class Movimiento
 
     public decimal? CostoUnitario { get; set; }
 
-    public int? Cantidad { get; set; }
+    public float? Cantidad { get; set; }
 
     public string? Motivo { get; set; }
